requirements.txt
# python==3.11.5
django==5.0.2
djangorestframework==3.14.0
django-filter==23.5